requirements.txt
aiogram<3.0
environs~=8.0.0
aiohttp~=3.8.1
ujson~=5.1.0
SQLAlchemy~=1.3.24
gino~=1.0.1
alembic~=1.7.7
pyzbar~=0.1.9
Pillow~=9.1.1
pyzbar==0.1.9
Pillow==9.1.1
opencv-python~=4.5.5.64
argparse~=1.4.0
numpy~=1.22.3